% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba.R
\name{rba_exchange_rates}
\alias{rba_exchange_rates}
\title{Fetch RBA exchange rate data (table F11)}
\usage{
rba_exchange_rates()
}
\value{
tibble: date, series_id, series_name, value
}
\description{
Convenience wrapper around rba_table("f11") for exchange rates.
}
