% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba.R
\name{rba_interest_rates}
\alias{rba_interest_rates}
\title{Fetch RBA interest rate data (table F01)}
\usage{
rba_interest_rates()
}
\value{
tibble: date, series_id, series_name, value
}
\description{
Convenience wrapper around rba_table("f01") for interest rates.
}
